+++ matchbox-frontend/src/app/validate/validation-entry.ts
@@ -1,5 +1,5 @@
-import { IssueSeverity, OperationResult } from '../util/operation-result';
-import {ValidationParameter, ValidationParameterDefinition} from "./validation-parameter";
+import {IssueSeverity, OperationResult} from '../util/operation-result';
+import {ValidationParameter} from "./validation-parameter";
 
 export class ValidationEntry {
   readonly filename: string; // "package/package.json",
@@ -9,14 +9,18 @@ export class ValidationEntry {
   readonly mimetype: string;
   result: OperationResult | undefined;
   aiRecommendation: string;
-  readonly profiles: string[] = [];
-  selectedProfile: string;
+  readonly extractedProfiles: string[] = [];
+  validationProfile: string;
   ig?: string;
   readonly date: Date;
   readonly validationParameters: ValidationParameter[] = [];
   public loading: boolean = false;
 
-  constructor(filename: string, resource: string, mimetype: string | null, profiles: string[] | null, settings: ValidationParameter[] = []) {
+  constructor(filename: string,
+              resource: string,
+              mimetype: string | null,
+              settings: ValidationParameter[] = [],
+              validationProfile: string | null = null) {
     this.filename = filename;
     this.resource = resource;
     this.validationParameters = settings;
@@ -31,10 +35,8 @@ export class ValidationEntry {
       }
     }
 
-    if (profiles) {
-      this.profiles = profiles;
-    }
     this.date = new Date();
+    this.validationProfile = validationProfile;
 
     try {
       if (this.mimetype === 'application/fhir+json') {
@@ -44,12 +46,6 @@ export class ValidationEntry {
       }
     } catch (e) {
       console.error('Error parsing resource to validate: ', e);
-    }
-
-    if (this.profiles && this.profiles.length) {
-      this.selectedProfile = this.profiles[0];
-    } else if (this.resourceType) {
-      this.selectedProfile = 'http://hl7.org/fhir/StructureDefinition/' + this.resourceType;
     }
   }
 
@@ -90,7 +86,7 @@ export class ValidationEntry {
       this.resourceId = res.id;
     }
     if (res.meta?.profile) {
-      this.profiles.push(...res.meta.profile);
+      this.extractedProfiles.push(...res.meta.profile);
     }
   }
 
@@ -115,7 +111,7 @@ export class ValidationEntry {
         let posProfileValue = this.resource.indexOf('value="', posProfileLeft) + 7;
         let posProfileValueRight = this.resource.indexOf('"', posProfileValue);
         if (posProfileValue < posProfileValueRight) {
-          this.profiles.push(this.resource.substring(posProfileValue, posProfileValueRight));
+          this.extractedProfiles.push(this.resource.substring(posProfileValue, posProfileValueRight));
         }
       }
     }